+++ packages/pn-commons/src/index.tsx
@@ -4,6 +4,7 @@ import Footer from './components/Footer/Footer';
 import Header from './components/Header/Header';
 import { LoadingOverlay } from './components/Loading/LoadingOverlay';
 import CustomPagination from './components/Pagination/CustomPagination';
+import CustomTooltip from './components/CustomTooltip';
 
 // types
 import { AppError } from './types/AppError';
@@ -15,6 +16,7 @@ export { Header };
 export { Layout };
 export { Footer };
 export { CustomPagination };
+export { CustomTooltip };
 
 // types
 export type { AppError };
