+++ packages/pn-pa-webapp/src/pages/components/Notifications/StatusTooltip.tsx
@@ -1,24 +1,5 @@
-import { useState } from 'react';
-
 import Chip from '@mui/material/Chip';
-import Tooltip, { TooltipProps, tooltipClasses } from '@mui/material/Tooltip';
-import ClickAwayListener from '@mui/material/ClickAwayListener';
-import { styled } from '@mui/material/styles';
-
-// TODO: utilizzare colori tema
-const BootstrapTooltip = styled(({ className, ...props }: TooltipProps) => (
-  <Tooltip {...props} arrow classes={{ popper: className }} />
-))(() => ({ // { theme }
-  [`& .${tooltipClasses.arrow}`]: {
-    color: '#455B71',
-  },
-  [`& .${tooltipClasses.tooltip}`]: {
-    backgroundColor: '#455B71'
-  },
-  [`& .${tooltipClasses.tooltip} .title`]: {
-    textAlign: 'center'
-  }
-}));
+import { CustomTooltip } from '@pagopa-pn/pn-commons';
 
 function StatusTooltip({tooltip, label, color}: {tooltip: string; label: string; color: "warning" | "error" | "success" | "info" | "default" | "primary" | "secondary" | undefined}) {
   const [title, body] = tooltip.split(':');
@@ -28,36 +9,11 @@ function StatusTooltip({tooltip, label, color}: {tooltip: string; label: string;
       <div className="title">{title.trim().toUpperCase()}</div>
       <div>{body.trim()}</div>
     </div>;
-  
-    // tooltip state
-    const [open, setOpen] = useState(false);
-    const handleTooltipClose = () => {
-      setOpen(false);
-    };
-    const handleTooltipOpen = () => {
-      setOpen(true);
-    };
 
   return (
-    <ClickAwayListener onClickAway={handleTooltipClose}>
-      <div>
-        <BootstrapTooltip  
-          title={tooltipContent}
-          arrow
-          placement="bottom"
-          PopperProps={{
-            disablePortal: true,
-          }}
-          onClose={handleTooltipClose}
-          open={open}
-          disableFocusListener
-          disableHoverListener
-          disableTouchListener
-        >
-          <Chip label={label} onClick={handleTooltipOpen} color={color}/>
-        </BootstrapTooltip >
-      </div>
-    </ClickAwayListener>
+    <CustomTooltip openOnClick tooltipContent={tooltipContent}>
+      <Chip label={label} color={color}/>
+    </CustomTooltip>
   );
 }
